WeatherForecastFirebase/requirerequirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
jws==0.1.3
oauth2client==4.1.3
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pycryptodomex==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
sweetify==2.0.0
urllib3==1.26.4
